main_worker/requirements.txt
celery[redis]
requests==2.11.1
beautifulsoup4==4.9.3
word-forms==2.1.0
Pillow==8.3.2
flower